+++ web/apps/photos/src/components/ExportFinished.tsx
@@ -19,7 +19,8 @@ interface Props {
     collectionNameMap: Map<number, string>;
     onHide: () => void;
     lastExportTime: number;
-    startExport: () => void;
+    /** Called when the user presses the "Resync" button. */
+    onResync: () => void;
 }
 
 export default function ExportFinished(props: Props) {
@@ -67,11 +68,7 @@ export default function ExportFinished(props: Props) {
                 <Button color="secondary" size="large" onClick={props.onHide}>
                     {t("CLOSE")}
                 </Button>
-                <Button
-                    size="large"
-                    color="primary"
-                    onClick={props.startExport}
-                >
+                <Button size="large" color="primary" onClick={props.onResync}>
                     {t("EXPORT_AGAIN")}
                 </Button>
             </DialogActions>