+++ web/apps/photos/src/services/deduplicationService.ts
@@ -3,7 +3,7 @@ import { FILE_TYPE } from "@/media/file-type";
 import type { Metadata } from "@/media/types/file";
 import { EnteFile } from "@/new/photos/types/file";
 import log from "@/next/log";
-import { apiOrigin } from "@/next/origins";
+import { apiURL } from "@/next/origins";
 import HTTPService from "@ente/shared/network/HTTPService";
 import { getToken } from "@ente/shared/storage/localStorage/helpers";
 
@@ -146,7 +146,7 @@ function groupDupesByFileHashes(dupe: Duplicate) {
 async function fetchDuplicateFileIDs() {
     try {
         const response = await HTTPService.get(
-            `${apiOrigin()}/files/duplicates`,
+            await apiURL("/files/duplicates"),
             null,
             {
                 "X-Auth-Token": getToken(),