+++ desktop/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v1.7.2 (Unreleased)
 
+-   Significantly improve the speed of the metadata parsing step during imports
+    of Google takeouts.
+-   Add a option to set and use a custom endpoint.
+-   Fix an issue preventing subscription purchases and renewals.
+-   Clear cached password after changing it on a different device.
+-   Reconcile exported files with disk on app start and resync.
 -   .
 
 ## v1.7.1